Moved Caching logic & data to `AutocompleteAdapter` class. - Moved original Adapter's `getSuggestions` to `getLSSuggestions` that queries the LS directly, and used `getSuggestions` - Tried to ensure minimal amount of breaking is done to the interface of the adapter..
