+++ lib/adapters/autocomplete-adapter.js
@@ -10,6 +10,12 @@ import {
   type ServerCapabilities,
 } from '../languageclient';
 import Convert from '../convert';
+import {filter} from 'fuzzaldrin-plus';
+
+type CompletionResponseCache = {
+  prefixPosition: atom$PointObject,
+  prefixes: Map<string, AutocompletionResponse>,
+};
 
 export type AutocompletionResponse = {
   isComplete: boolean,
@@ -19,6 +25,8 @@ export type AutocompletionResponse = {
 // Public: Adapts the language server protocol "textDocument/completion" to the Atom
 // AutoComplete+ package.
 export default class AutocompleteAdapter {
+  _completionCachesByEditor: Map<atom$TextEditor, CompletionResponseCache> = new Map();
+
   static canAdapt(serverCapabilities: ServerCapabilities): boolean {
     return serverCapabilities.completionProvider != null;
   }
@@ -32,6 +40,144 @@ export default class AutocompleteAdapter {
   // Returns a {Promise} of an {AutocompletionResponse} object containing the AutoComplete+
   // suggestions to display and an indication if the list is complete.
   async getSuggestions(
+    connection: LanguageClientConnection,
+    request: atom$AutocompleteRequest,
+    useCache: boolean = false,
+    triggerChars?: Array<string>,
+  ): Promise<AutocompletionResponse> {
+    if (!useCache) {
+      return await this.getLSSuggestions(connection, request);
+    }
+
+    if (!triggerChars || triggerChars.length === 0) {
+      throw Error('Must provide triggerChars to get suggestions with caching');
+    }
+
+    let prefixPosition: atom$PointObject;
+
+    // Calculating the potential position of a trigger character.
+    // If the current prefix is a trigger character we use it's position, otherwise -
+    // we take the position request.prefix.length chars backwards.
+    if (triggerChars.includes(request.prefix)) {
+      prefixPosition = request.bufferPosition;
+    } else {
+      prefixPosition = {
+        row: request.bufferPosition.row,
+        column: request.bufferPosition.column - request.prefix.length,
+      };
+    }
+
+    const triggerChar: string = request.editor
+      .getBuffer()
+      .getTextInRange([[prefixPosition.row, prefixPosition.column - 1], prefixPosition]);
+
+    if (!triggerChars.includes(triggerChar)) {
+      return Promise.resolve({isComplete: true, completionItems: []});
+    }
+
+    // Basically, we're going to build a completion cache for each editor in the form of an object
+    // that stores the last prefix position from which we fetched results from the LS,
+    // and a Map of results for each known prefix in that position.
+    // Everytime we hit the LanguageServer for suggestions, we're going to store those results
+    // in the cache.
+    let completionCache: ?CompletionResponseCache = this._completionCachesByEditor.get(request.editor);
+    if (completionCache == null) {
+      completionCache = {
+        prefixPosition,
+        prefixes: new Map(),
+      };
+    }
+
+    // Checking if the current trigger character is positioned where we stored
+    // our cache on the last fetch from the LS.
+    if (
+      completionCache.prefixPosition.row === prefixPosition.row &&
+      completionCache.prefixPosition.column === prefixPosition.column &&
+      completionCache.prefixes &&
+      completionCache.prefixes.size > 0
+    ) {
+      let isFilterRequired: boolean;
+      let resultsForPrefix: ?AutocompletionResponse;
+      let k: number = request.prefix.length;
+
+      // We try to get cached results for the longest possible substring of our
+      // prefix. Everytime we fail to, we search on a shorter substring (going backwards).
+      // If we find cached results, but they are incomplete - we break the loop - we must
+      // fetch results from the language server.
+      while (!resultsForPrefix && k >= 0) {
+        let _resultsForPrefix: ?AutocompletionResponse;
+        if (k === 0) {
+          _resultsForPrefix = completionCache.prefixes.get(triggerChar);
+        } else {
+          _resultsForPrefix = completionCache.prefixes.get(request.prefix.slice(0, k));
+        }
+
+        if (_resultsForPrefix) {
+          // If we found results but they are incomplete - there's no point to keep searching -
+          // the smaller the query substring is, the more results we're going to get.
+          if (_resultsForPrefix.isComplete) {
+            resultsForPrefix = _resultsForPrefix;
+            isFilterRequired = k !== request.prefix.length;
+          }
+
+          break;
+        }
+
+        k--;
+      }
+
+      if (resultsForPrefix && resultsForPrefix.completionItems) {
+        return new Promise(resolve => {
+          const listToFilter = (resultsForPrefix && resultsForPrefix.completionItems) || [];
+          const filtered = isFilterRequired
+            ? filter(listToFilter, request.prefix, {key: 'text'}).map(s => ({
+                ...s,
+                ...{replacementPrefix: request.prefix},
+              }))
+            : listToFilter;
+          resolve({
+            isComplete: true,
+            completionItems: filtered,
+          });
+        });
+      }
+    }
+
+    // If the prefix position of the current request is different than the last
+    // cached one, we re-instantiate the completion cache with the new position
+    // and no results.
+    if (
+      completionCache.prefixPosition.row !== prefixPosition.row ||
+      completionCache.prefixPosition.column !== prefixPosition.column
+    ) {
+      completionCache = {
+        prefixPosition,
+        prefixes: new Map(),
+      };
+    }
+
+    const lsResponse = await this.getLSSuggestions(connection, request);
+    if (!completionCache) {
+      completionCache = {
+        prefixPosition,
+        prefixes: new Map(),
+      };
+    }
+
+    completionCache.prefixes.set(request.prefix, lsResponse);
+    this._completionCachesByEditor.set(request.editor, completionCache);
+    return lsResponse;
+  }
+
+  // Public: Primary entry point for obtaining suggestions for AutoComplete+ by
+  // querying the language server.
+  //
+  // * `connection` A {LanguageClientConnection} to the language server to query.
+  // * `request` An {Object} with the AutoComplete+ request to satisfy.
+  //
+  // Returns a {Promise} of an {AutocompletionResponse} object containing the AutoComplete+
+  // suggestions to display and an indication if the list is complete.
+  async getLSSuggestions(
     connection: LanguageClientConnection,
     request: atom$AutocompleteRequest,
   ): Promise<AutocompletionResponse> {
