+++ lib/auto-languageclient.js
@@ -7,10 +7,9 @@ import {CompositeDisposable, Disposable} from 'atom';
 import {ConsoleLogger, NullLogger, type Logger} from './logger';
 import {ServerManager, type ActiveServer} from './server-manager.js';
 import Convert from './convert.js';
-import {filter} from 'fuzzaldrin-plus';
 
 import ApplyEditAdapter from './adapters/apply-edit-adapter';
-import AutocompleteAdapter, {type AutocompletionResponse} from './adapters/autocomplete-adapter';
+import AutocompleteAdapter from './adapters/autocomplete-adapter';
 import CodeActionAdapter from './adapters/code-action-adapter';
 import CodeFormatAdapter from './adapters/code-format-adapter';
 import CodeHighlightAdapter from './adapters/code-highlight-adapter';
@@ -25,11 +24,6 @@ import SignatureHelpAdapter from './adapters/signature-help-adapter';
 
 type ConnectionType = 'stdio' | 'socket' | 'ipc';
 
-type CompletionResponseCache = {
-  prefixPosition: atom$PointObject,
-  prefixes: Map<string, AutocompletionResponse>,
-};
-
 // Public: AutoLanguageClient provides a simple way to have all the supported
 // Atom-IDE services wired up entirely for you by just subclassing it and
 // implementing startServerProcess/getGrammarScopes/getLanguageName and
@@ -39,7 +33,6 @@ export default class AutoLanguageClient {
   _serverManager: ServerManager;
   _linterDelegate: linter$V2IndieDelegate;
   _signatureHelpRegistry: ?atomIde$SignatureHelpRegistry;
-  _completionCachesByEditor: Map<atom$TextEditor, CompletionResponseCache> = new Map();
 
   processStdErr: string = '';
   logger: Logger;
@@ -271,7 +264,7 @@ export default class AutoLanguageClient {
 
     if (request.activatedManually) {
       this.autoComplete = this.autoComplete || new AutocompleteAdapter();
-      return this.autoComplete.getSuggestions(server.connection, request).then(response => {
+      return this.autoComplete.getLSSuggestions(server.connection, request).then(response => {
         return response.completionItems;
       });
     }
@@ -288,116 +281,8 @@ export default class AutoLanguageClient {
       triggerChars = ['.'];
     }
 
-    let prefixPosition: atom$PointObject;
-
-    // Calculating the potential position of a trigger character.
-    // If the current prefix is a trigger character we use it's position, otherwise -
-    // we take the position request.prefix.length chars backwards.
-    if (triggerChars.includes(request.prefix)) {
-      prefixPosition = request.bufferPosition;
-    } else {
-      prefixPosition = {
-        row: request.bufferPosition.row,
-        column: request.bufferPosition.column - request.prefix.length,
-      };
-    }
-
-    const triggerChar: string = request.editor
-      .getBuffer()
-      .getTextInRange([[prefixPosition.row, prefixPosition.column - 1], prefixPosition]);
-
-    if (!triggerChars.includes(triggerChar)) {
-      return Promise.resolve([]);
-    }
-
-    // Basically, we're going to build a completion cache for each editor in the form of an object
-    // that stores the last prefix position from which we fetched results from the LS,
-    // and a Map of results for each known prefix in that position.
-    // Everytime we hit the LanguageServer for suggestions, we're going to store those results
-    // in the cache.
-    let completionCache: ?CompletionResponseCache = this._completionCachesByEditor.get(request.editor);
-    if (completionCache == null) {
-      completionCache = {
-        prefixPosition,
-        prefixes: new Map(),
-      };
-    }
-
-    // Checking if the current trigger character is positioned where we stored
-    // our cache on the last fetch from the LS.
-    if (
-      completionCache.prefixPosition.row === prefixPosition.row &&
-      completionCache.prefixPosition.column === prefixPosition.column &&
-      completionCache.prefixes &&
-      completionCache.prefixes.size > 0
-    ) {
-      let isFilterRequired: boolean;
-      let resultsForPrefix: ?AutocompletionResponse;
-      let k: number = request.prefix.length;
-
-      // We try to get cached results for the longest possible substring of our
-      // prefix. Everytime we fail to, we search on a shorter substring (going backwards).
-      // If we find cached results, but they are incomplete - we break the loop - we must
-      // fetch results from the language server.
-      while (!resultsForPrefix && k >= 0) {
-        let _resultsForPrefix: ?AutocompletionResponse;
-        if (k === 0) {
-          _resultsForPrefix = completionCache.prefixes.get(triggerChar);
-        } else {
-          _resultsForPrefix = completionCache.prefixes.get(request.prefix.slice(0, k));
-        }
-
-        if (_resultsForPrefix) {
-          // If we found results but they are incomplete - there's no point to keep searching -
-          // the smaller the query substring is, the more results we're going to get.
-          if (_resultsForPrefix.isComplete) {
-            resultsForPrefix = _resultsForPrefix;
-            isFilterRequired = k !== request.prefix.length;
-          }
-
-          break;
-        }
-
-        k--;
-      }
-
-      if (resultsForPrefix && resultsForPrefix.completionItems) {
-        return new Promise(resolve => {
-          const listToFilter = (resultsForPrefix && resultsForPrefix.completionItems) || [];
-          const filtered = isFilterRequired
-            ? filter(listToFilter, request.prefix, {key: 'text'}).map(s => ({
-                ...s,
-                ...{replacementPrefix: request.prefix},
-              }))
-            : listToFilter;
-          resolve(filtered);
-        });
-      }
-    }
-
-    // If the prefix position of the current request is different than the last
-    // cached one, we re-instantiate the completion cache with the new position
-    // and no results.
-    if (
-      completionCache.prefixPosition.row !== prefixPosition.row ||
-      completionCache.prefixPosition.column !== prefixPosition.column
-    ) {
-      completionCache = {
-        prefixPosition,
-        prefixes: new Map(),
-      };
-    }
-
     this.autoComplete = this.autoComplete || new AutocompleteAdapter();
-    return this.autoComplete.getSuggestions(server.connection, request).then(response => {
-      if (!completionCache) {
-        completionCache = {
-          prefixPosition,
-          prefixes: new Map(),
-        };
-      }
-      completionCache.prefixes.set(request.prefix, response);
-      this._completionCachesByEditor.set(request.editor, completionCache);
+    return this.autoComplete.getSuggestions(server.connection, request, true, triggerChars).then(response => {
       return response.completionItems;
     });
   }
